Backend/requirement.txt
fastapi
uvicorn
python-dotenv
motor
python-jose
passlib
bcrypt==4.0.1